Updated C++ standard and improved code readability and type safety
Updated the C++ language standard from C++17 to C++20 across multiple files. Improved code readability in `FFTCtrl.cpp` and ensured correct data type for variables across various files. Removed unnecessary condition in `rdptrs[0] != theAudioBuffer->getReadPointer(0)`. Changed `deltaHz` variable type from `double` to `float`. Explicitly cast several variables to their correct types using `static_cast`. Initialized `fftData` array with zeros. Commented out unused parameters and functions.

+++ Source/AudioPlaybackModule.h
@@ -276,7 +276,7 @@ public:
 			[this]
 			{
 				//audioSourcePlayer.setGain(gainSlider.getValue());
-				transportSource.setGain(gainSlider.getValue());
+				transportSource.setGain(static_cast<float>(gainSlider.getValue()));
 			};
 		gainSlider.setValue(1.0f);
